server/requirements.txt
python-dotenv==0.13.0
Django==3.0.6
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-oauth-toolkit==1.3.2
django-cors-headers==3.2.1
django-apscheduler==0.3.0
APScheduler==3.6.3
requests==2.23.0
pillow==7.1.2  # for model image field
google-auth==1.14.3
google-api-python-client==1.8.3